Rename config.h to board_config.h
